feat(handdecomp): carve PutWMFaceOnBg — DECL_ONLY + cast-hoist + permuter (+1 → 7769)
D204: three stacked JP diffs, each needed for diff 0:
1. DECL_ONLY: declare GetWMFaceTileXAndScroll (JP-bound, // ??? stub).
2. cast-hoist: 5th param flip is s8 (entry asrs); `int f = flip` forces the
   single entry sign-ext + drops the redundant re-extension (136→132B).
3. permuter (base 20 → 0 in 9078 iters) localized a callee-proto bug:
   PutFaceOnBackGround's 3rd param is int in JP, not fe8u's s8 — the s8
   narrowing at the call pinned an r9<->sl swap. Declaring it `int c`
   (zero blast-radius: callee still asm, only src/ caller is PutWMFaceOnBg)
   flips it. Full cold make compare OK, self-contain 100%.

+++ include/face.h
@@ -152,7 +152,7 @@ void FaceChange_Finish(struct UnkFaceProc* proc);
 void StartFaceChange(struct FaceProc* parent, int fid);
 void DecompressFaceImg(int offset, int fid);
 void ApplyFacePalette(int pal, int fid);
-void PutFaceOnBackGround(u16* a, u16 b, s8 c);
+void PutFaceOnBackGround(u16* a, u16 b, int c);
 
 extern const struct FaceData portrait_data[];
 extern struct FaceProc* gFaces[];
